[bugfix] alpha blending wasn't (quite) right
Anyway, this might or might not have been a bug, but seeing and
understanding the code myself, I see that you're not supposed to use a
16-bit value but an 8-bit one.

Correcting this doesnt do anything for some reason, but for the hell of
it, I'm considering that it's "fixed".

+++ src/output/graphical_wayland/render.h
@@ -8,7 +8,7 @@
 
 #include "main.h"
 
-extern uint32_t wayland_color_blend (uint32_t color, uint16_t alpha);
+extern uint32_t wayland_color_blend (uint32_t color, uint8_t alpha);
 extern void     update_frame        (struct waydata *s);
 extern void     reallocSHM          (struct waydata *wd);
 extern void     closeSHM            (struct waydata *wd);
